+++ radix-engine/src/engine/mod.rs
@@ -1,27 +1,29 @@
+mod actor;
 mod auth_module;
 mod call_frame;
 mod errors;
+mod execution_state;
+mod kernel;
+mod node;
+mod node_properties;
+mod node_ref;
 mod precommitted_kv_store;
-mod re_execution_state;
-mod re_kernel;
-mod re_node;
 mod system_api;
 mod track;
 mod track_support;
-mod type_properties;
-mod values;
 mod wasm_runtime;
 
+pub use actor::*;
 pub use auth_module::*;
 pub use call_frame::CallFrame;
 pub use errors::*;
+pub use execution_state::*;
+pub use kernel::*;
+pub use node::*;
+pub use node_properties::*;
+pub use node_ref::*;
 pub use precommitted_kv_store::*;
-pub use re_execution_state::*;
-pub use re_kernel::*;
-pub use re_node::*;
 pub use system_api::SystemApi;
 pub use track::*;
 pub use track_support::*;
-pub use type_properties::*;
-pub use values::*;
 pub use wasm_runtime::*;
